+++ WhiteQueen.java
@@ -1,14 +1,14 @@
 
 public class WhiteQueen extends Gamepiece{
 
-	public final int val;
+	public final byte val;
 	
-	public WhiteQueen(int x, int y) {
+	public WhiteQueen(byte x, byte y) {
 		super(x, y);
 		val = 1;
 	}
 	
-	public WhiteQueen(char x, int y) {
+	public WhiteQueen(char x, byte y) {
 		super(x, y);
 		val = 1;
 	}
@@ -22,23 +22,23 @@ public class WhiteQueen extends Gamepiece{
 		return super.position();
 	}
 	
-	public int val() {
+	public byte val() {
 		return val;
 	}
 	
-	public void move(int x, int y) {
+	public void move(byte x, byte y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public void move(char x, int y) {
-		this.x = (x > 96 ? x - 97 : x - 65);
+	public void move(char x, byte y) {
+		this.x = (byte)(x > 96 ? x - 97 : x - 65);
 		this.y = y;
 	}
 	
 	public void move(String c) {
 		int to_x = c.charAt(0);
-		this.x = (to_x > 96 ? to_x - 97 : to_x - 65);
-		this.y = Integer.parseInt(c.substring(1)) - 1;
+		this.x = (byte)(to_x > 96 ? to_x - 97 : to_x - 65);
+		this.y = (byte)(Integer.parseInt(c.substring(1)) - 1);
 	}
 }
